Вынес функцию в отдельный файл
